% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa.R
\name{gpa}
\alias{gpa}
\title{Calculate GPA from numeric grades.}
\usage{
gpa(x, ID = ID, numericgrade = numericgrade, Unit.Taken = Unit.Taken)
}
\arguments{
\item{x}{Dataframe containing student-course data including ID and numericgrades.}

\item{ID}{EmplID variable name.}

\item{numericgrade}{Name for columns containing grades in numeric format.}

\item{Unit.Taken}{Name for column representing the number of units for a course in integer format.
ID refers to student EmplID. Numericgrades can be created with gradetonumgrade()
Note that this does not account for credits at this time.}
}
\description{
Calculate GPA from numeric grades.
}
\details{
Used for student-course data e.g. one row per course per student.
Assumes that the grades have already been converted to numeric values.
Calculates the GPA for and the number of courses the GPA is based on
Attempts to handle replaced grades and the CUNY F and CUNY D policy correctly.
}

