% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradetonumgrade.R
\name{gradetonumgrade}
\alias{gradetonumgrade}
\title{Function to convert letter grades to numbers.}
\usage{
gradetonumgrade(x)
}
\arguments{
\item{x}{is a dataframe containing the Grade field from CUNYFirst}
}
\description{
Function to convert letter grades to numbers.
}

