% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.peoplesoft.csv.R
\name{read.peoplesoft.csv}
\alias{read.peoplesoft.csv}
\title{Function to read a downloaded csv file.}
\usage{
read.peoplesoft.csv(filename)
}
\arguments{
\item{filename}{full name and location of the csv file to be read.}
}
\description{
Function to read a downloaded csv file.
}
\details{
Skips the extraneous line in the last row that is added by CUNYFirst.
}

